+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog 1.0.0].
 
+## Unpublished
+
+- **Reparse/Enrichment**: Don't reparse/enrich documents that have external data
+
 ## v40.0.0 (2025-08-20)
 
 ### Feat